pyproject.toml
[project]
name = "blichat-bot"
version = "0.1.0"
description = "B站直播弹幕与Telegram双向互动机器人"
requires-python = ">=3.10"
dependencies = [
    "blivedm @ git+https://github.com/xfgryujk/blivedm.git",
    "python-telegram-bot>=21.0",
    "bilibili-api-python>=16.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
